Switch to glfw 3.4

+++ UI/Examples/example_glfw_opengl3/timer.h
@@ -31,7 +31,7 @@ public:
             endTime = m_EndTime;
         }
 
-        return std::chrono::duration_cast<std::chrono::milliseconds>(endTime - m_StartTime).count();
+        return static_cast<double>(std::chrono::duration_cast<std::chrono::milliseconds>(endTime - m_StartTime).count());
     }
 
     double elapsedSeconds()
